requirements.txt
fastapi
uvicorn
slack-sdk
openai
sqlite3
python-dotenv
